pyproject.toml
[build-system]
requires = [
    "setuptools>=65",
]
build-backend = "setuptools.build_meta"

[project]
name = "mcschematic-plus"
version = "0.0.1"
description = "Generate Minecraft schem and nbt structures from 3D image data"
readme = "README.md"
requires-python = ">=3.9,<3.12"
license = {file = "LICENSE"}
authors = [
    {name = "Kevin Tan"},
]
classifiers = [
    "Programming Language :: Python :: 3",
]

dependencies = [
    "numpy",
    "tifffile",
    "pyvista",
    "mcschematic",
    "scipy",
    "pillow",
]

[project.optional-dependencies]
mesh_to_sdf = ["trimesh", "mesh_to_sdf"]
amulet = ["amulet-core"]
jupyter = ["jupyterlab", "pyvista[jupyter]"]
all = ["mcschematic-plus[mesh_to_sdf, amulet, jupyter]"]

[tool.setuptools.package-data]
mcschematic_plus = ["data/block_colormaps/*.csv"]